code review changes

+++ services/resource-usage-tracker/src/simcore_service_resource_usage_tracker/models/resource_tracker_pricing_details.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 
 from models_library.resource_tracker import PricingDetailId, PricingPlanId
 from pydantic import BaseModel
@@ -8,7 +9,7 @@ class PricingDetailDB(BaseModel):
     pricing_detail_id: PricingDetailId
     pricing_plan_id: PricingPlanId
     unit_name: str
-    cost_per_unit: float
+    cost_per_unit: Decimal
     valid_from: datetime
     valid_to: datetime | None
     specific_info: dict
